+++ docs/changes.rst
@@ -18,6 +18,10 @@ Changelog
 - Updated Patternslib to 9.9.1, fixing pat-display-time issue
   [reinhardt]
 
+- Allow activating consultancy per country.
+  Ref: `scrum-1265 <https://github.com/syslabcom/scrum/issues/1265>`_
+  [reinhardt]
+
 - Don't show consultant role if consultancy phase is disabled.
   Ref: `scrum-1225 <https://github.com/syslabcom/scrum/issues/1225>`_
   [reinhardt]